+++ src/ui.h
@@ -1,0 +1,8 @@
+#include <pebble.h>
+
+void init_ui(ClickHandler on_select_click);
+void cleanup_ui();
+
+void set_text(char* text);
+
+void scroll_to_top();
